[Feature] Created Filter Object
Since the raw sensor data is noisy, it is highly desirable to filter
the output to get a useful signal.  This commit introduces a filter
class that includes a running average filter as the default.  There
is an Enum which selects could be used to select other types of
filters in future commits.